third_party/WebKit/LayoutTests/storage/websql/fts-crash-703507.html

<!doctype html>
<meta charset="utf-8">
<title>WebSQL: FTS3 Crash in bug 703507</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';

async_test(testCase => {
  const database = openDatabase(
      'Fts3CrashTest', '1.0', 'Database for FTS3 crash test', 1024 * 1024);
  assert_not_equals(database, null, 'openDatabase should not fail');

  database.transaction(testCase.step_func(transaction => {
    transaction.executeSql(
        'DROP TABLE IF EXISTS main;', [], () => {},
        testCase.unreached_func('FTS3 table drop should not fail'));
    transaction.executeSql(
        'CREATE VIRTUAL TABLE main USING FTS3(fts3table);', [],
        testCase.unreached_func('FTS3 table creation should not succeed'),
        () => { testCase.done() });
  }));

}, 'The FTS3 extension should not be available');

</script>
